reword `toList_` docs
As a beginner the original wording didn't convey what the real issue was and scared me off.  "Space leak" is easily interpreted as something much more serious than it really is.  Now that I know better I'll fix it for the next beginner. Ref https://neilmitchell.blogspot.com/2008/12/mapm-mapm-and-monadic-statements.html